+++ src/Component/Result/Debug/Detail.php
@@ -12,7 +12,7 @@ namespace Solarium\Component\Result\Debug;
 /**
  * Select component debug detail result.
  */
-class Detail
+class Detail implements \ArrayAccess
 {
     /**
      * Value.
@@ -85,22 +85,60 @@ class Detail
     }
 
     /**
-     * @param array $subDetails
+     * @param \Solarium\Component\Result\Debug\Detail[]|array $subDetails
      *
      * @return self
      */
     public function setSubDetails(array $subDetails): self
     {
-        $this->subDetails = $subDetails;
+        $this->subDetails = [];
+        foreach ($subDetails as $subDetail) {
+            $this->subDetails[] = new Detail($subDetail['match'], $subDetail['value'], $subDetail['description']);
+        }
 
         return $this;
     }
 
     /**
-     * @return array
+     * @return \Solarium\Component\Result\Debug\Detail[]|null
      */
-    public function getSubDetails(): array
+    public function getSubDetails(): ?array
     {
         return $this->subDetails;
+    }
+
+    public function offsetExists($offset)
+    {
+        return in_array($offset, ['match', 'value', 'description']);
+    }
+
+    public function offsetGet($offset)
+    {
+        return $this->{$offset};
+    }
+
+    public function offsetSet($offset, $value)
+    {
+        // Details are immutable.
+    }
+
+    public function offsetUnset($offset)
+    {
+        // Details are immutable.
+    }
+
+    public function __toString()
+    {
+        $string = '';
+        if ($this->match) {
+            $string .= sprintf('%f', $this->value).' <= '.$this->description.PHP_EOL;
+            foreach ($this->getSubDetails() as $subDetail) {
+                if ($subDetail->getMatch()) {
+                    $string .= '... '.$subDetail;
+                }
+            }
+        }
+
+        return $string;
     }
 }