+++ src/Component/Result/Debug/Document.php
@@ -24,7 +24,7 @@ class Document extends Detail implements \IteratorAggregate, \Countable
     /**
      * Details.
      *
-     * @var array
+     * @var \Solarium\Component\Result\Debug\Detail[]
      */
     protected $details;
 
@@ -57,7 +57,7 @@ class Document extends Detail implements \IteratorAggregate, \Countable
     /**
      * Get details.
      *
-     * @return array
+     * @return \Solarium\Component\Result\Debug\Detail[]
      */
     public function getDetails(): array
     {
@@ -82,5 +82,15 @@ class Document extends Detail implements \IteratorAggregate, \Countable
     public function count(): int
     {
         return \count($this->details);
+    }
+
+    public function __toString()
+    {
+        $string = '';
+        foreach ($this->getDetails() as $detail) {
+            $string .= '  '.$detail.PHP_EOL;
+        }
+
+        return $string;
     }
 }